+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 All notable changes to this project will be documented in this file.
 
 
-## [0.1.5] - 2020-12-23
+## [0.2.0] - 2020-12-23
 ### Added:
 - 'format_t' class
 - 'format::riff' class
-- Default RIFF chunk reader (skip)
+- Default RIFF chunk handler (copy)
+- WAVE chunk handlers (fmt/data)